+++ lib/src/lib.rs
@@ -53,6 +53,7 @@ pub mod lock;
 pub mod matchers;
 pub mod merge;
 pub mod merged_tree;
+pub mod mock_signing;
 pub mod object_id;
 pub mod op_heads_store;
 pub mod op_store;